requirements-dev.txt
# =============================================================================
# CSAT-Compass — Ontwikkelafhankelijkheden
# Installeer met: pip install -r requirements-dev.txt
# Bevat runtime-deps (via requirements.txt) + IDE/test-tools
# =============================================================================

# Alle runtime-afhankelijkheden inbegrepen
-r requirements.txt

# --- Type stubs (IDE code insight) ---
pandas-stubs>=3.0.0

# --- Linting & formatting ---
ruff>=0.15.0
mypy>=1.9.0

# --- Security ---
bandit>=1.8.0
pip-audit>=2.9.0

# --- Pre-commit ---
pre-commit>=4.0.0

# --- Testing ---
pytest>=7.0
pytest-cov>=4.0

